+++ config.py
@@ -4,19 +4,19 @@ import torch
 
 from albumentations.pytorch import ToTensorV2
 
-DATASET = 'PASCAL_VOC'
-DEVICE = 'cuda' if torch.cuda.is_available() else 'cpu'
+DATASET = "PASCAL_VOC"
+DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
 NUM_WORKERS = 2
 BATCH_SIZE = 32
 IMAGE_SIZE = 416
 NUM_CLASSES = 20
-LEARNING_RATE = 1e-5
+LEARNING_RATE = 1e-4
 WEIGHT_DECAY = 1e-4
 NUM_EPOCHS = 100
-CONF_THRESHOLD = 0.05 # probability of object being detected is greater than this
+CONF_THRESHOLD = 0.6  # probability of object being detected is greater than this
 MAP_IOU_THRESH = 0.5
 NMS_IOU_THRESH = 0.45
-S = [IMAGE_SIZE // 32, IMAGE_SIZE // 16, IMAGE_SIZE // 8] # grid size
+S = [IMAGE_SIZE // 32, IMAGE_SIZE // 16, IMAGE_SIZE // 8]  # grid size 13, 26, 52
 PIN_MEMORY = True
 LOAD_MODEL = False
 SAVE_MODEL = True
@@ -28,10 +28,10 @@ ANCHORS = [
     [(0.28, 0.22), (0.38, 0.48), (0.9, 0.78)],
     [(0.07, 0.15), (0.15, 0.11), (0.14, 0.29)],
     [(0.02, 0.03), (0.04, 0.07), (0.08, 0.06)],
-]  
+]
 # These values have been normalized
 
-scale = 1.1
+scale = 1.2
 train_transforms = A.Compose(
     [
         A.LongestMaxSize(max_size=int(IMAGE_SIZE * scale)),
@@ -57,10 +57,18 @@ train_transforms = A.Compose(
         A.Posterize(p=0.1),
         A.ToGray(p=0.1),
         A.ChannelShuffle(p=0.05),
-        A.Normalize(mean=[0, 0, 0], std=[1, 1, 1], max_pixel_value=255,),
+        A.Normalize(
+            mean=[0, 0, 0],
+            std=[1, 1, 1],
+            max_pixel_value=255,
+        ),
         ToTensorV2(),
     ],
-    bbox_params=A.BboxParams(format="yolo", min_visibility=0.4, label_fields=[],),
+    bbox_params=A.BboxParams(
+        format="yolo",
+        min_visibility=0.4,
+        label_fields=[],
+    ),
 )
 
 test_transforms = A.Compose(
@@ -69,7 +77,11 @@ test_transforms = A.Compose(
         A.PadIfNeeded(
             min_height=IMAGE_SIZE, min_width=IMAGE_SIZE, border_mode=cv2.BORDER_CONSTANT
         ),
-        A.Normalize(mean=[0, 0, 0], std=[1, 1, 1], max_pixel_value=255,),
+        A.Normalize(
+            mean=[0, 0, 0],
+            std=[1, 1, 1],
+            max_pixel_value=255,
+        ),
         ToTensorV2(),
     ],
     bbox_params=A.BboxParams(format="yolo", min_visibility=0.4, label_fields=[]),
@@ -95,87 +107,88 @@ PASCAL_CLASSES = [
     "sheep",
     "sofa",
     "train",
-    "tvmonitor"
+    "tvmonitor",
 ]
 
-COCO_LABELS = ['person',
- 'bicycle',
- 'car',
- 'motorcycle',
- 'airplane',
- 'bus',
- 'train',
- 'truck',
- 'boat',
- 'traffic light',
- 'fire hydrant',
- 'stop sign',
- 'parking meter',
- 'bench',
- 'bird',
- 'cat',
- 'dog',
- 'horse',
- 'sheep',
- 'cow',
- 'elephant',
- 'bear',
- 'zebra',
- 'giraffe',
- 'backpack',
- 'umbrella',
- 'handbag',
- 'tie',
- 'suitcase',
- 'frisbee',
- 'skis',
- 'snowboard',
- 'sports ball',
- 'kite',
- 'baseball bat',
- 'baseball glove',
- 'skateboard',
- 'surfboard',
- 'tennis racket',
- 'bottle',
- 'wine glass',
- 'cup',
- 'fork',
- 'knife',
- 'spoon',
- 'bowl',
- 'banana',
- 'apple',
- 'sandwich',
- 'orange',
- 'broccoli',
- 'carrot',
- 'hot dog',
- 'pizza',
- 'donut',
- 'cake',
- 'chair',
- 'couch',
- 'potted plant',
- 'bed',
- 'dining table',
- 'toilet',
- 'tv',
- 'laptop',
- 'mouse',
- 'remote',
- 'keyboard',
- 'cell phone',
- 'microwave',
- 'oven',
- 'toaster',
- 'sink',
- 'refrigerator',
- 'book',
- 'clock',
- 'vase',
- 'scissors',
- 'teddy bear',
- 'hair drier',
- 'toothbrush'
+COCO_LABELS = [
+    "person",
+    "bicycle",
+    "car",
+    "motorcycle",
+    "airplane",
+    "bus",
+    "train",
+    "truck",
+    "boat",
+    "traffic light",
+    "fire hydrant",
+    "stop sign",
+    "parking meter",
+    "bench",
+    "bird",
+    "cat",
+    "dog",
+    "horse",
+    "sheep",
+    "cow",
+    "elephant",
+    "bear",
+    "zebra",
+    "giraffe",
+    "backpack",
+    "umbrella",
+    "handbag",
+    "tie",
+    "suitcase",
+    "frisbee",
+    "skis",
+    "snowboard",
+    "sports ball",
+    "kite",
+    "baseball bat",
+    "baseball glove",
+    "skateboard",
+    "surfboard",
+    "tennis racket",
+    "bottle",
+    "wine glass",
+    "cup",
+    "fork",
+    "knife",
+    "spoon",
+    "bowl",
+    "banana",
+    "apple",
+    "sandwich",
+    "orange",
+    "broccoli",
+    "carrot",
+    "hot dog",
+    "pizza",
+    "donut",
+    "cake",
+    "chair",
+    "couch",
+    "potted plant",
+    "bed",
+    "dining table",
+    "toilet",
+    "tv",
+    "laptop",
+    "mouse",
+    "remote",
+    "keyboard",
+    "cell phone",
+    "microwave",
+    "oven",
+    "toaster",
+    "sink",
+    "refrigerator",
+    "book",
+    "clock",
+    "vase",
+    "scissors",
+    "teddy bear",
+    "hair drier",
+    "toothbrush",
 ]
